pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
    "flask-session>=0.6.0",
    "chromadb>=1.1.0",
    "cryptography>=46.0.0",
    "joblib>=1.4.2",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.1",
    "shap>=0.46.0",
    "sqlalchemy>=2.0.0",
    "pytest>=8.0.0",
    "xgboost>=2.1.1",
    "langgraph>=1.1.3",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langchain-community>=0.4.1",
    "pymupdf>=1.24.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "frontend*", "migrations*", "*"]
exclude = ["venv*", "tests*"]
